feat: Delete jobs S3 object

+++ lib/api/jobs.ts
@@ -4,7 +4,7 @@ import { JobKind, Jobs } from "../db/schema/cam";
 import zod, { object } from "zod";
 import { createSelectSchema } from "drizzle-zod";
 import { client } from "../aws";
-import { PutObjectCommand } from "@aws-sdk/client-s3";
+import { DeleteObjectCommand, PutObjectCommand } from "@aws-sdk/client-s3";
 import { scopeNames as scopes } from "../scopes";
 
 const RequestSchema = createSelectSchema(Jobs).pick({ kind: true, payload: true });
@@ -49,7 +49,13 @@ export const Complete = routeFactory(async (req, authType, tx, id) => {
 export const DELETE = routeFactory(async (req, authType, tx, id) => {
   if (!id) return routeResponse(422);
   const job = await tx.query.Jobs.findFirst({ where: eq(Jobs.id, id) });
-  await checkUserTeam(tx, authType, job?.team_id);
+  if (!job) return routeResponse(404);
+  await checkUserTeam(tx, authType, job.team_id);
 
-  return await tx.delete(Jobs).where(eq(Jobs.id, id)).returning({ id: Jobs.id });
+  const result = await tx.delete(Jobs).where(eq(Jobs.id, id)).returning({ id: Jobs.id });
+  await client.send(new DeleteObjectCommand({
+    Bucket: process.env.AUTOCAM_BUCKET,
+    Key: `teams/${job.team_id}/jobs/${id}`
+  }));
+  return result;
 }, { emailVerifiedNeeded: true, requiredScopes: [scopes.jobs.delete] });
